Arithmetic operation (15 point(s))
Write a program to realize the basic arithmetic operation. Input two numbers and an operator(‘+’, ‘-’ , ‘’, ‘/’), calculate the result and output it.

Input Specification:
Input two floating point numbers and an operator seperated by space, the order of imput is: number1 operator number2

Output Specification:
If the operator is one of the '+', '-', '*', '/' and the result can be calculated correctly : Output the fomula and the result, the numbers and operators are seperated by space, and all the numbers remain two digits after the decimal point. Otherwise, output "Error!"

Sample Input:
Write a sample input here. For example:

3.4 + 5.2
Sample Output:
Write the corresponding sample output here. For example:

3.40 + 5.20 = 8.60